refactor lib to be more dev friendly
