+++ experiments/common-wasm/build.sbt
@@ -5,7 +5,7 @@ lazy val scala213 = "2.13.11"
 lazy val supportedScalaVersions = List(scala212, scala213)
 
 ThisBuild / scalaVersion     := scala212
-ThisBuild / version          := "1.0.0-SNAPSHOT"
+ThisBuild / version          := "16.8.0-dev"
 ThisBuild / organization     := "io.otoroshi.common"
 ThisBuild / organizationName := "wasm"
