+++ experiments/common-wasm/build.sh
@@ -1,3 +1,4 @@
+rm ./target/scala-2.12/common-wasm_2.12-*.jar
 sbt '+package'
-rm ../../otoroshi/lib/common-wasm_2.12-1.0.0-SNAPSHOT.jar
-cp ./target/scala-2.12/common-wasm_2.12-1.0.0-SNAPSHOT.jar ../../otoroshi/lib/
+rm ../../otoroshi/lib/common-wasm_2.12-*.jar
+cp ./target/scala-2.12/common-wasm_2.12-*.jar ../../otoroshi/lib/
