+++ otoroshi/app/auth/wasm.scala
@@ -1,16 +1,15 @@
 package otoroshi.auth
 
-import io.otoroshi.common.wasm.WasmFunctionParameters
 import otoroshi.env.Env
 import otoroshi.gateway.Errors
 import otoroshi.models._
 import otoroshi.next.models.NgRoute
 import otoroshi.next.plugins.BodyHelper
-import otoroshi.next.plugins.api.{NgAccess, NgCachedConfigContext}
+import otoroshi.next.plugins.api.NgCachedConfigContext
 import otoroshi.next.utils.JsonHelpers
 import otoroshi.security.IdGenerator
-import otoroshi.utils.{JsonPathValidator, TypedMap}
 import otoroshi.utils.syntax.implicits._
+import otoroshi.utils.{JsonPathValidator, TypedMap}
 import play.api.Logger
 import play.api.libs.json._
 import play.api.mvc._
@@ -142,14 +141,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
               maybeRoute = ctx.route.some
             )
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("pa_login_page", input.stringify), None)
+          vm.callExtismFunction("pa_login_page", input.stringify)
             .map {
               case Left(err)     => Results.InternalServerError(err)
               case Right(output) => {
                 val response    =
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
@@ -211,14 +209,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
             )
             .map(_.left)
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("pa_logout", input.stringify), None)
+          vm.callExtismFunction("pa_logout", input.stringify)
             .map {
               case Left(err)     => Results.InternalServerError(err).left
               case Right(output) => {
                 val response  =
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
@@ -258,14 +255,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
       env.wasmIntegration.wasmVmFor(plugin.config).flatMap {
         case None          => "plugin not found !".leftf
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("pa_callback", input.stringify), None)
+          vm.callExtismFunction("pa_callback", input.stringify)
             .map {
               case Left(err)     => err.stringify.left
               case Right(output) => {
                 val response = {
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
@@ -310,14 +306,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
               maybeRoute = ctx.route.some
             )
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("bo_login_page", input.stringify), None)
+          vm.callExtismFunction("bo_login_page", input.stringify)
             .map {
               case Left(err)     => Results.InternalServerError(err)
               case Right(output) => {
                 val response    =
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
@@ -374,14 +369,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
             )
             .map(_.left)
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("bo_logout", input.stringify), None)
+          vm.callExtismFunction("bo_logout", input.stringify)
             .map {
               case Left(err)     => Results.InternalServerError(err).left
               case Right(output) => {
                 val response  =
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
@@ -418,14 +412,13 @@ class WasmAuthModule(val authConfig: WasmAuthModuleConfig) extends AuthModule {
       env.wasmIntegration.wasmVmFor(plugin.config).flatMap {
         case None          => "plugin not found !".leftf
         case Some((vm, _)) =>
-          implicit val ctx = env.wasmIntegrationCtx
-          vm.call(WasmFunctionParameters.ExtismFuntionCall("bo_callback", input.stringify), None)
+          vm.callExtismFunction("bo_callback", input.stringify)
             .map {
               case Left(err)     => err.stringify.left
               case Right(output) => {
                 val response = {
                   try {
-                    Json.parse(output._1)
+                    Json.parse(output)
                   } catch {
                     case e: Exception =>
                       WasmAuthModule.logger.error("error during json parsing", e)
