+++ otoroshi/app/controllers/BackOfficeController.scala
@@ -10,7 +10,7 @@ import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
 import com.google.common.base.Charsets
 import com.nimbusds.jose.jwk.KeyType
-import io.otoroshi.common.wasm.WasmManagerSettings
+import io.otoroshi.common.wasm.scaladsl._
 import org.joda.time.DateTime
 import org.mindrot.jbcrypt.BCrypt
 import org.slf4j.LoggerFactory
